refactor(subway): 예외 목록 패키지 분할
