+++ src/main/java/subway/exception/line/LineNameFormatException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.line;
 
 public class LineNameFormatException extends RuntimeException {
 }
