+++ src/main/java/subway/exception/line/LineNameLengthException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.line;
 
 public class LineNameLengthException extends RuntimeException {
 }
