+++ src/main/java/subway/exception/station/SameUpLineAndDownLineException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class SameUpLineAndDownLineException extends RuntimeException {
 }
