+++ src/main/java/subway/exception/station/StationAlreadyExistException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class StationAlreadyExistException extends RuntimeException {
 }
