+++ src/main/java/subway/exception/station/StationNameFormatException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class StationNameFormatException extends RuntimeException {
 }
