+++ src/main/java/subway/exception/station/StationNameLengthException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class StationNameLengthException extends RuntimeException {
 }
