+++ src/main/java/subway/exception/station/StationRegisteredOnLineException.java
@@ -1,4 +1,4 @@
-package subway.exception;
+package subway.exception.station;
 
 public class StationRegisteredOnLineException extends RuntimeException {
 }
